requirements.txt
fastmcp
groq
fastapi
uvicorn
requests
python-dotenv
chromadb
sentence-transformers
